jiangjiang_entry/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.4.1
async-timeout==4.0.1
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.9
django-heroku==0.3.1
django-import-export==2.7.0
et-xmlfile==1.1.0
frozenlist==1.2.0
future==0.18.2
gunicorn==20.1.0
idna==3.3
line-bot-sdk==2.0.1
MarkupPy==1.14
multidict==5.2.0
odfpy==1.4.1
openpyxl==3.0.9
psycopg2==2.9.3
python-decouple==3.5
pytz==2021.3
PyYAML==6.0
requests==2.26.0
sqlparse==0.4.2
static3==0.7.0
tablib==3.1.0
typing_extensions==4.0.0
urllib3==1.26.7
whitenoise==5.3.0
xlrd==2.0.1
xlwt==1.3.0
yarl==1.7.2
